Added useIsDesktop hook in skills and navbar
