+++ src/hooks/useIsDesktop.tsx
@@ -1,0 +1,18 @@
+import { useEffect, useState } from "react";
+import { breakpoints } from "../styles/variables";
+
+export default function useIsDesktop() {
+  const [isDesktop, setIsDesktop] = useState(false);
+
+  useEffect(() => {
+    function handleResize() {
+      setIsDesktop(window.innerWidth > breakpoints.md);
+    }
+    handleResize();
+    window.addEventListener("resize", handleResize);
+
+    return () => window.removeEventListener("resize", handleResize);
+  }, [breakpoints]);
+
+  return isDesktop;
+}
